Duckdb

+++ tutake/api/ts/index_stock.py
@@ -7,30 +7,32 @@ Tushare index_weight接口
 
 @author: rmfish
 """
+from datetime import datetime, timedelta
+
 import pandas as pd
 import requests
 from bs4 import BeautifulSoup
 from sqlalchemy import Integer, String, Column
 from sqlalchemy.orm import sessionmaker
 
-from tutake.api.base_dao import Base, BatchWriter
+from tutake.api.base_dao import Base, BatchWriter, BaseDao, TutakeTableBase
 from tutake.api.process import DataProcess
+from tutake.api.process_client import Task
 from tutake.api.ts.tushare_api import TushareAPI
 from tutake.api.ts.tushare_dao import TushareDAO, create_shared_engine
 from tutake.utils.config import TutakeConfig
 from tutake.utils.utils import project_root
 
 
-class TushareIndexStock(Base):
+class TushareIndexStock(TutakeTableBase):
     __tablename__ = "tushare_index_stock"
-    id = Column(Integer, primary_key=True, autoincrement=True)
     index_code = Column(String, index=True, comment='指数代码')
     con_code = Column(String, index=True, comment='成分代码')
     list_date = Column(String, index=True, comment='纳入日期')
     delist_date = Column(String, comment='剔除日期')
 
 
-class IndexStock(TushareDAO, DataProcess):
+class IndexStock(TushareDAO, Task, DataProcess):
     instance = None
 
     def __new__(cls, *args, **kwargs):
@@ -40,8 +42,9 @@ class IndexStock(TushareDAO, DataProcess):
 
     def __init__(self, config):
         self.table_name = "tushare_index_stock"
-        self.database = 'tushare_index.db'
-        self.database_url = config.get_data_sqlite_driver_url(self.database)
+        self.database = 'tutake.duckdb'
+        self.database_dir = config.get_tutake_data_dir()
+        self.database_url = config.get_data_driver_url(self.database)
         self.engine = create_shared_engine(self.database_url,
                                            connect_args={
                                                'check_same_thread': False,
@@ -50,6 +53,7 @@ class IndexStock(TushareDAO, DataProcess):
         session_factory = sessionmaker()
         session_factory.configure(bind=self.engine)
         TushareIndexStock.__table__.create(bind=self.engine, checkfirst=True)
+        self.schema = BaseDao.parquet_schema(TushareIndexStock)
 
         query_fields = ['index_code', 'con_code', 'start_date', 'end_date', 'limit', 'offset']
         self.tushare_fields = ["index_code", "con_code", "list_date", "delist_date"]
@@ -57,7 +61,8 @@ class IndexStock(TushareDAO, DataProcess):
         column_mapping = None
         TushareDAO.__init__(self, self.engine, session_factory, TushareIndexStock, self.database, self.table_name,
                             query_fields, entity_fields, column_mapping, config)
-        DataProcess.__init__(self, "index_weight", config)
+        DataProcess.__init__(self, "index_stock", config)
+        Task.__init__(self, "index_stock", config)
         self.api = TushareAPI(config)
 
     def columns_meta(self):
@@ -104,7 +109,8 @@ class IndexStock(TushareDAO, DataProcess):
         同步历史数据
         :return:
         """
-        return super()._process(self.fetch_and_append, BatchWriter(self.engine, self.table_name), **kwargs)
+        return super()._process(self.fetch_and_append,
+                                BatchWriter(self.engine, self.table_name, self.schema, self.database_dir), **kwargs)
 
     def _index_stock_cons(self, index_code: str = "000300.SH", latest=True) -> pd.DataFrame:
         """
@@ -190,6 +196,8 @@ class IndexStock(TushareDAO, DataProcess):
         csi100:000903.SH
         csi50:000016.SH
 
+
+        000009,000016,000010
         index_code	display_name	publish_date
         000001	上证指数	1991/7/15
         000002	A股指数	1992/2/21
@@ -874,12 +882,31 @@ class IndexStock(TushareDAO, DataProcess):
         399996	中证智能家居指数	2014/9/17
         399997	中证白酒指数	2015/1/21
         399998	中证煤炭指数	2015/2/13
+        000035 000032  399170
         """
-        return [{"index_code": "000300.SH"}, {"index_code": "000905.SH"}, {"index_code": "000852.SH"},
-                {"index_code": "000903.SH"}, {"index_code": "000016.SH"}]
 
-    def prepare(self):
-        self.delete_all()
+        return [{"index_code": "000300.SH"}, {"index_code": "000905.SH"}, {"index_code": "000852.SH"},
+                {"index_code": "000903.SH"}, {"index_code": "000016.SH"}, {"index_code": "000037.SH"},
+                {"index_code": "000039.SH"}, {"index_code": "000035.SH"}, {"index_code": "000032.SH"},
+                {"index_code": "399170.SH"}]
+
+    def need_to_process(self, **kwargs) -> bool:
+        size = self.select("count(distinct(index_code))")
+        if size != len(kwargs['params']):
+            return True
+        max_date = self.max("list_date")
+        if max_date is not None:
+            return max_date < (datetime.now() + timedelta(days=-30)).strftime("%Y%m%d")
+        return True
+
+    def prepare_write(self, writer, **kwargs) -> bool:
+        db_cnt = self.count()
+        w_cnt = writer.count()
+        if w_cnt >= db_cnt:
+            self.delete_all()
+            return True
+        else:
+            return False
 
     def default_limit(self) -> str:
         return "10000"
